#90 Done Small Changes
Fixed the Labels for the threshold
Fixed the Labels for the line style

Still need to add Line style for Thresholds.
Still need to add color picker for line of best fit and thresholds.

Still need to fix the line of best fit, it is not fixed when swapping axes.