Update scatter.py
fixed when choose any axes to be logarithmic, the best fitness line will be turn off and disabled
fixed the equation shown in the legend to be 3 significant figures
fixed the chosen color could be saved now
fixed the markers could be changed seperatly and saved, same as the color

Furthermore, I think all of the users stores of scatter have been done, I do not have more time to do extra things than what the users stores asked. Demas and Robin's formats are good, but I have to move on to the other two projects, wchich are dued on next Friday, time is not enough. Thanks for understanding.