Update scatter.py
fixed when chooses any axes to be logarithmic, the line of best fit will be turned off and disabled
fixed the equation legend to be rounded to 3 significant figures
fixed the colors could be saved for each chosen value
fixed the markers could be saved for each chosen value, same as the color
Furthermore, I think all of the users storys have been completed, and I do not have more time to extra things, Demas and Robin's styles are pretty good, but unfortunately, I have to move on to the other projects, which are dued on next Friday. Thanks for understanding.